requirements.txt
black>=20.8b1
codecov>=2.0.15
hypothesis==6.9.2
matplotlib>=3.1.1
networkx>=2.3
numpy>=1.17.3
pillow>=7.2.0
pydantic>=1.8.1
pylint>=2.3.1
pytest-black>=0.3.11
pytest-cov>=2.7.1
pytest-isort>=1.2.0
pytest-mypy>=0.3.3
pytest-pylint>=0.17.0
pytest>=5.0.1
rply>=0.7.7